feat: Add Categories component for displaying product categories
This commit adds a new component called Categories, which is responsible for displaying a list of product categories. The component uses data from a JSON file and renders each category as a card with an image and name. The Categories component is designed to be scrollable horizontally and has a responsive layout.